+++ entity/__init__.py
@@ -1,11 +1,13 @@
+from entity.bbridge_entity import BBridgeEntity
+from entity.concept_detection_data import ConceptDetectionData
 from entity.image_concepts import ImagesConcepts, Concepts
 from entity.image_objects import ImageObjects, Object
-from entity.concept_detection_data import ConceptDetectionData
-from entity.object_detection_data import ObjectDetectionData
 from entity.ner import NER, Entity
 from entity.nlp_data import NLPData
+from entity.object_detection_data import ObjectDetectionData
 from entity.pos_tagging import POSTagging, POS
-from entity.response import RequestId, Response
+from entity.request_id import RequestId
+from entity.response import Response
 from entity.sentiments import Sentiments
 from entity.user import User
 from entity.user_profile import UserAttributes, UserProfile