+++ entity/object_detection_data.py
@@ -1,7 +1,7 @@
-import json
+from entity import BBridgeEntity
 
 
-class ObjectDetectionData(object):
+class ObjectDetectionData(BBridgeEntity):
     def __init__(self, url, threshold):
         """
         :type url: str
@@ -10,6 +10,10 @@ class ObjectDetectionData(object):
         self.__url = url
         self.__threshold = threshold
 
+    @classmethod
+    def from_json(cls, json_object):
+        return ObjectDetectionData(json_object["url"], json_object["threshold"])
+
     @property
     def url(self):
         return self.__url
@@ -17,19 +21,3 @@ class ObjectDetectionData(object):
     @property
     def threshold(self):
         return self.__threshold
-
-    @staticmethod
-    def from_json(json_object):
-        """
-        :type json_object: dict
-        :rtype: entity.object_detection_data.ObjectDetectionData
-        """
-        return ObjectDetectionData(json_object["url"], json_object["threshold"])
-
-    @staticmethod
-    def from_json_str(json_string):
-        """
-        :type json_string: str
-        :rtype: entity.object_detection_data.ObjectDetectionData
-        """
-        return ObjectDetectionData.from_json(json.loads(json_string))