+++ entity/pos_tagging.py
@@ -1,36 +1,24 @@
-import json
+from entity import BBridgeEntity
 
 
-class POSTagging(object):
+class POSTagging(BBridgeEntity):
     def __init__(self, results):
         """
         :type results: list of (list of entity.pos_tagging.POS)
         """
         self.__results = results
 
+    @classmethod
+    def from_json(cls, json_object):
+        results = [[POS.from_json(x) for x in xs] for xs in json_object["results"]]
+        return POSTagging(results)
+
     @property
     def results(self):
         return self.__results
 
-    @staticmethod
-    def from_json(json_object):
-        """
-        :type json_object: dict
-        :rtype: entity.pos_tagging.POSTagging
-        """
-        results = [[POS.from_json(x) for x in xs] for xs in json_object["results"]]
-        return POSTagging(results)
 
-    @staticmethod
-    def from_json_str(json_string):
-        """
-        :type json_string: str
-        :rtype: entity.pos_tagging.POSTagging
-        """
-        return POSTagging.from_json(json.loads(json_string))
-
-
-class POS(object):
+class POS(BBridgeEntity):
     def __init__(self, text, type_):
         """
         :type text: str
@@ -39,6 +27,10 @@ class POS(object):
         self.__text = text
         self.__type = type_
 
+    @classmethod
+    def from_json(cls, json_object):
+        return POS(json_object["text"], json_object["type"])
+
     @property
     def text(self):
         return self.__text
@@ -46,19 +38,3 @@ class POS(object):
     @property
     def type(self):
         return self.__type
-
-    @staticmethod
-    def from_json(json_object):
-        """
-        :type json_object: dict
-        :rtype: entity.pos_tagging.POS
-        """
-        return POS(json_object["text"], json_object["type"])
-
-    @staticmethod
-    def from_json_str(json_string):
-        """
-        :type json_string: str
-        :rtype: entity.pos_tagging.POS
-        """
-        return POS.from_json(json.loads(json_string))