+++ entity/response.py
@@ -1,10 +1,7 @@
-import json
-
-
 class Response(object):
     def __init__(self, body, status_code=200, err_message=None):
         """
-        :type body: object | None
+        :type body: entity.bbridge_entity.BBridgeEntity | dict | None
         :type status_code: int
         :type err_message: str | None
         """
@@ -23,31 +20,3 @@ class Response(object):
     @property
     def err_message(self):
         return self.__err_message
-
-
-class RequestId(object):
-    def __init__(self, request_id):
-        """
-        :type request_id: str
-        """
-        self.__request_id = request_id
-
-    @property
-    def request_id(self):
-        return self.__request_id
-
-    @staticmethod
-    def from_json(json_object):
-        """
-        :type json_object: dict
-        :rtype: entity.response.RequestId
-        """
-        return RequestId(json_object["request_id"])
-
-    @staticmethod
-    def from_json_str(json_string):
-        """
-        :type json_string: str
-        :rtype: entity.response.RequestId
-        """
-        return RequestId.from_json(json.loads(json_string))