+++ CHANGELOG.md
@@ -15,6 +15,9 @@
     * *ImageConcepts* → *Concepts*
     * *ImageObject* → *Object*
 
+### Features
+
+* All bBridge entity classes inherit BBridgeEntity class
 
 # v0.1 (2017-03-03)